[usb][dwc3] Move dwc3 driver and tests to GN
Migrate the DWC3 USB controller driver and its tests from Bazel to the
GN build system to enable code coverage data collection.

Test: dwc3-test
Depends-on: turquoise-internal:Ibe7e061f641538e8b6ef9adcd4d89d4718876641
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1642358
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Dependencies-Satisfied: cl-deps-checker@fuchsia-internal-service-accts.iam.gserviceaccount.com <cl-deps-checker@fuchsia-internal-service-accts.iam.gserviceaccount.com>

+++ src/devices/usb/drivers/dwc3/dwc3-commands.cc
@@ -130,7 +130,7 @@ void Dwc3::CmdEpEndTransfer(const Endpoint& ep) {
 
   auto* mmio = get_mmio();
 
-  const uint32_t ep_num = ep.ep_num;
+  const uint8_t ep_num = ep.ep_num;
   const uint32_t rsrc_id = ep.rsrc_id;
 
   ZX_DEBUG_ASSERT_MSG(rsrc_id != Endpoint::kInvalidResourceId,
@@ -165,7 +165,7 @@ void Dwc3::CmdEpSetStall(const Endpoint& ep) {
   TRACE_DURATION("dwc3", "Dwc3::CmdEpSetStall", "ep_num", ep.ep_num);
   auto* mmio = get_mmio();
 
-  const uint32_t ep_num = ep.ep_num;
+  const uint8_t ep_num = ep.ep_num;
 
   DEPCMDPAR0::Get(ep_num).FromValue(0).WriteTo(mmio);
   DEPCMDPAR1::Get(ep_num).FromValue(0).WriteTo(mmio);
@@ -184,7 +184,7 @@ void Dwc3::CmdEpClearStall(const Endpoint& ep) {
   TRACE_DURATION("dwc3", "Dwc3::CmdEpClearStall", "ep_num", ep.ep_num);
   auto* mmio = get_mmio();
 
-  const uint32_t ep_num = ep.ep_num;
+  const uint8_t ep_num = ep.ep_num;
 
   DEPCMDPAR0::Get(ep_num).FromValue(0).WriteTo(mmio);
   DEPCMDPAR1::Get(ep_num).FromValue(0).WriteTo(mmio);
